Do not install storks-scheduler for non-PX backups

+++ test/integration_test/test-deploy.sh
@@ -146,22 +146,24 @@ for i in $(seq 1 100) ; do
     fi
 done
 
-echo "Creating stork scheduler"
-kubectl apply -f /specs/stork-scheduler.yaml
+if [ "$volume_driver" == "pxd" ] ; then
+	echo "Creating stork scheduler"
+	kubectl apply -f /specs/stork-scheduler.yaml
 # Delete the pods to make sure we are waiting for the status from the
 # new pods
-kubectl delete pods -n kube-system -l name=stork-scheduler
+	kubectl delete pods -n kube-system -l name=stork-scheduler
 
-echo "Waiting for stork-scheduler to be in running state"
-for i in $(seq 1 100) ; do
-    replicas=$(kubectl get deployment -n kube-system stork-scheduler -o json | jq ".status.readyReplicas")
-    if [ "$replicas" == "3" ]; then
-        break
-    else
-        echo "Stork scheduler is not ready yet"
-        sleep 10
-    fi
-done
+	echo "Waiting for stork-scheduler to be in running state"
+	for i in $(seq 1 100) ; do
+		replicas=$(kubectl get deployment -n kube-system stork-scheduler -o json | jq ".status.readyReplicas")
+		if [ "$replicas" == "3" ]; then
+			break
+		else
+			echo "Stork scheduler is not ready yet"
+			sleep 10
+		fi
+	done
+fi
 
 if [ "$run_cluster_domain_test" == "true" ] ; then
 	sed -i 's/'enable_cluster_domain'/'\""true"\"'/g' /testspecs/stork-test-pod.yaml
